chore: moving componet function to a separeted file
